Add _ALWAYS versions of TEST_MSG and TEST_DUMP macros
As part of using acutest for CI testing, it can be useful to record
additional data beyond a simple success / failure.

The addition of _ALWAYS versions of TEST_MSG AND TEST_DUMP macro allows
for extra data to be captured during the running of tests which is then
ouptput to stdout.

Also, data captured by use of TEST_MSG* and TEST_DUMP* macros is stored
in log entries which are passed from child test processes to the parent
to then be included in the XML output produced when the -x / --xml-output
option is used.

The output is presented in <system-out> nodes within each <testcase>
node as per the jUnit XSD specification.